Fixed SCons scripts to build with unitTest++. Therefore removed the Makefile that had to be used to compile tddTest project.

+++ usr/share/tddTest/main.cpp
@@ -35,8 +35,8 @@ unsigned int nat(int v1)
 
 TEST(nat)
 {
-    CHECK_EQUAL(nat(2), 2);
-    CHECK_EQUAL(nat(1), 1);
+    CHECK_EQUAL(nat(2), (unsigned int)2);
+    CHECK_EQUAL(nat(1), (unsigned int)1);
     CHECK_THROW(nat(-1), std::string);
     CHECK_THROW(nat(0), std::string);
